+++ source/sequencer/MidiSequence.c
@@ -48,7 +48,8 @@ void appendMidiEventToSequence(MidiSequence midiSequence, MidiEvent midiEvent) {
   }
 }
 
-boolByte fillMidiEventsFromRange(MidiSequence midiSequence, const unsigned long startTimestamp, const int blocksize, LinkedList outMidiEvents) {
+boolByte fillMidiEventsFromRange(MidiSequence midiSequence, const unsigned long startTimestamp,
+  const unsigned long blocksize, LinkedList outMidiEvents) {
   MidiEvent midiEvent;
   LinkedListIterator iterator = midiSequence->_lastEvent;
   const unsigned long stopTimestamp = startTimestamp + blocksize;
